third_party/WebKit/LayoutTests/editing/style/style-3998892-fix.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
    '<div contenteditable><div>foo^</div><div>bar|</div></div>',
    'bold',
    '<div contenteditable><div>foo^</div><div><b>bar|</b></div></div>'),
    'Bolding a selection from end of line should not change unselected text on starting line.');
</script>
